requirements.txt
PySide6==6.7.2
PyYAML>=6.0
